refactor: refactored the existing code
