+++ src/components/Header/Header.jsx
@@ -3,27 +3,15 @@ import { Link } from "react-router-dom";
 import { AiOutlineSetting } from "react-icons/ai";
 import { BsSunFill } from "react-icons/bs";
 import { GiToggles } from "react-icons/gi";
-import { GrFormClose } from "react-icons/gr";
 import { FiSearch } from "react-icons/fi";
-import { useState } from "react";
+import { useHeader } from "../../contexts/header-context";
+import { SearchBar } from "../SearchBar/SearchBar";
+import { FilterAndSortPopUp } from "../FilterAndSortPopUp/FilterAndSortPopUp";
 
 const Header = () => {
 
-    const [ popUpFilterAndSortDisplay, setPopUpFilterAndSortDisplay ] = useState("none");
-    const [ popUpSearchBarDisplay, setPopUpSearchBarDisplay ] = useState("none");
-
-    const getThePopUpSearchBarDisplay = () => {
-        popUpSearchBarDisplay === "none" ? setPopUpSearchBarDisplay("flex") : setPopUpSearchBarDisplay("none")
-    }
-
-    const getThePopUpFilterAndSortDisplay =() => {
-        popUpFilterAndSortDisplay === "none" ? setPopUpFilterAndSortDisplay("flex") : setPopUpFilterAndSortDisplay("none")
-    }
-
-    const closeThePopUpFilterAndSort = () => {
-        setPopUpFilterAndSortDisplay("none")
-    }
-
+    const { getThePopUpSearchBarDisplay, getThePopUpFilterAndSortDisplay } = useHeader();
+    
     return (
         <div className="tn_header-component">
             <div className="tn_header-content">
@@ -40,54 +28,9 @@ const Header = () => {
                 </div>
             </div>
 
-            <div className="tn_searchBar-popUp-outerContainer popUp-outerContainer" style={{ display: popUpSearchBarDisplay }}>
-                <div className="tn_searchBar-popUp popUp b-rad1">
-                    <input type="text" placeholder="Search" className="tn_searchBar b-rad1" />
-                </div>
-            </div>
+            <SearchBar />
 
-            <div className="tn_filter-popUp-outerContainer popUp-outerContainer" style={{ display: popUpFilterAndSortDisplay }}>
-                <div className="tn_filter-popUp-container popUp b-rad1">
-                    <div className="tn_filter-subcontainer1">
-                        <h4>Sort & Filter Notes</h4>
-                        <GrFormClose className="tn_filter-popUp-closeBtn" onClick={() => closeThePopUpFilterAndSort()} />
-                    </div>
-                    <div className="tn_filter-subcontainer2">
-
-                        <label htmlFor="tn_sort" className="tn_popUp-subheading" >Sort By</label>
-                        <input type="text" name="sortNotes" defaultValue="Newest First" list="sortList" id="tn_sort" />
-                        <datalist id="sortList">
-                            <option value="Newest First" />
-                        </datalist>
-
-                        <label htmlFor="tn_filter" className="tn_popUp-subheading" >Filter By</label>
-                        <input type="text" name="filterNotes" defaultValue="Labels" list="filterList" id="tn_filter" />
-                        <datalist id="filterList">
-                            <option value="Labels" />
-                            <option value="Pinned" />
-                            <option value="Archived" />
-                        </datalist>
-
-                        <div className="tn_popUp-subheading" >Select Labels</div>
-                        <label className="tn_select-label" htmlFor="label-1">
-                            <input type="checkbox" id="label-1" />
-                            Label 1
-                        </label>
-                        <label className="tn_select-label" htmlFor="label-2">
-                            <input type="checkbox" id="label-2" />
-                            Label 2
-                        </label>
-                        <label className="tn_select-label" htmlFor="label-3">
-                            <input type="checkbox" id="label-3" />
-                            Label 3
-                        </label>
-
-                    </div>
-                    <div className="tn_filter-subcontainer3">
-                        <button className="tn_popUp-done-btn et_p-simple-btn primary-color btn" onClick={() => closeThePopUpFilterAndSort()} >Done</button>
-                    </div>
-                </div>
-            </div>
+            <FilterAndSortPopUp />
 
         </div>
     );
