+++ src/components/Header/header.css
@@ -1,7 +1,9 @@
 @import url(../../utils/base.css);
 
 .tn_header-component {
-    box-shadow: 0px 0px 1px 1px var(--most-lightest-brown);
+    background-image: url(https://images.unsplash.com/photo-1511285605577-4d62fb50d2f7?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=876&q=80);
+    background-size: cover;
+    box-shadow: 0px 0px 3px 0px var(--black);
     position: relative;
 }
 
@@ -13,13 +15,13 @@
 }
 
 .tn_header-heading {
-    color: var(--light-gray);
+    color: var(--white-color);
     text-decoration: none;
 }
 
 .tn_header-searchBar {
     border: transparent;
-    box-shadow: 0px 0px 4px 2px var(--most-lightest-brown);
+    box-shadow: 0px 0px 4px 0px var(--light-gray);
     min-width: 30rem;
     outline: none;
     padding: 0.5rem 1rem;
@@ -37,122 +39,19 @@
 }
 
 .tn_header-profile-icon {
+    border: 2px solid var(--white-color);
     margin: 0;
 }
 
 .header-icon {
     cursor: pointer;
-    color: var(--light-gray);
+    color: var(--white-color);
     font-size: 1.5rem;
-}
-
-.popUp-outerContainer {
-    display: flex;
-    justify-content: center;
-}
-
-.popUp {
-    background-color: var(--white-color);
-    border: 1px solid var(--most-lightest-brown);
-    box-shadow: 0px 0px 2px 0px var(--most-lightest-brown);
-    display: flex;
-    flex-direction: column;
-    gap: 0.5rem;
-    margin-top: 2rem;
-    position: absolute;
-    width: 22rem;
-    z-index: 2;
-}
-
-.tn_searchBar-popUp {
-    padding: 3rem 1rem;
-    width: 70%;
-}
-
-.tn_searchBar {
-    border: transparent;
-    box-shadow: 0px 0px 4px 2px var(--most-lightest-brown);
-    outline: none;
-    padding: 0.5rem 1rem;
-    margin: auto;
-    width: 80%;
-}
-
-.tn_filter-popUp-container>* {
-    display: flex;
-    padding: 1rem;
-}
-
-.tn_filter-subcontainer1 {
-    align-items: center;
-    border-bottom: 1px solid var(--most-lightest-brown);
-    justify-content: space-between;
-}
-
-.tn_filter-subcontainer1>* {
-    color: var(--light-gray);
-}
-
-.tn_filter-popUp-closeBtn {
-    cursor: pointer;
-    font-size: 1.3rem;
-}
-
-.tn_filter-subcontainer2 {
-    flex-direction: column;
-    justify-content: start;
-}
-
-.tn_popUp-subheading {
-    color: var(--light-gray);
-    font-size: small;
-    font-weight: bold;
-    padding-bottom: 0.5rem;
-    text-align: left;
-}
-
-.tn_filter-subcontainer2 input {
-    background-color: var(--light-brown);
-    border: 1px solid var(--most-lightest-brown);
-    color: var(--light-gray);
-    padding: 0.5rem 1rem;
-    margin-bottom: 1rem;
-    outline: none;
-}
-
-.tn_filter-subcontainer2 label {
-    color: var(--light-gray);
-    cursor: pointer;
-    display: flex;
-    gap: 0.5rem;
-    font-size: small;
-    text-align: left;
-}
-
-.tn_select-label input {
-    margin-top: 1px;
-    width: 20px;
-}
-
-.tn_filter-subcontainer3 {
-    display: flex;
-    justify-content: flex-end;
-}
-
-.tn_popUp-done-btn {
-    background-color: var(--light-brown);
-    color: var(--light-gray);
 }
 
 @media only screen and (max-width: 860px) {
     .tn_header-searchBar {
-        min-width: 23rem;
-    }
-}
-
-@media only screen and (max-width: 750px) {
-    .tn_header-searchBar {
-        min-width: 19rem;
+        min-width: 18rem;
     }
 }
 
@@ -172,17 +71,13 @@
     }
 }
 
-@media only screen and (max-width: 475px) {
+@media only screen and (max-width: 480px) {
     .tn_header-searchBar {
         display: none;
     }
 
     .tn_header-search-icon {
         display: block;
-    }
-
-    .tn_searchBar-popUp {
-        width: 90%;
     }
 }
