+++ src/components/SearchBar/searchBar.css
@@ -1,0 +1,15 @@
+@import url(../../utils/base.css);
+
+.tn_searchBar-popUp {
+    padding: 3rem 1rem;
+    width: 80%;
+}
+
+.tn_searchBar {
+    border: transparent;
+    box-shadow: 0px 0px 4px 0px var(--light-gray);
+    outline: none;
+    padding: 0.5rem 1rem;
+    margin: auto;
+    width: 90%;
+}
