+++ src/index.js
@@ -4,6 +4,8 @@ import "./index.css";
 import App from "./App";
 import { makeServer } from "./server";
 import { BrowserRouter } from "react-router-dom";
+import { NoteProvider } from "./contexts/note-context";
+import { HeaderProvider } from "./contexts/header-context";
 
 // Call make Server
 makeServer();
@@ -11,7 +13,11 @@ makeServer();
 ReactDOM.render(
   <StrictMode>
     <BrowserRouter>
-      <App />
+      <HeaderProvider>
+        <NoteProvider>
+          <App />
+        </NoteProvider>
+      </HeaderProvider>
     </BrowserRouter>
   </StrictMode>,
   document.getElementById("root")
